Add command to delete an issue
