+++ .github/workflows/generate-js.yml
@@ -48,7 +48,4 @@ jobs:
           branch: github-actions/generate-js/${{ github.ref_name }}
           title: Generate JavaScript for branch `${{ github.ref_name }}`
           body: |
-            This pull pequest is opened automatically and contains generated JavaScript code based on commit ${{ github.sha }}. The changes are the following (context excluded):
-            ```
-            ${{ steps.diff.outputs.GIT_DIFF }}
-            ```
+            This pull pequest is opened automatically and contains generated JavaScript code based on commit ${{ github.sha }}.
